refactor: Use Row struct to render styles better
The previous method of manipulating the cursor and filtered matches did
not allow for proper styling. This large refactor allows the easier
manipulation of an individual row as it now stores its own state.

Multiple table functions are also added to generate output for its rows
in multiple states - aligned and styled, aligned but not styled, plain
string.